+++ src/tootstream/toot.py
@@ -1251,27 +1251,42 @@ def unboost(mastodon, rest):
     cprint(msg, attr("dim"))
 
 
-@command("<id>", "Toots")
+@command("<id> [<id>]", "Toots")
 def fav(mastodon, rest):
-    """Favorites a toot by ID."""
-    rest = IDS.to_global(rest)
-    if rest is None:
-        return
-    faved = mastodon.status_favourite(rest)
-    msg = "  Favorited:\n" + get_content(faved)
-    cprint(msg, attr("dim"))
+    """Favorites a toot by ID or IDs."""
+    for i in range(0, 2):
+        rest = rest.replace("  ", " ")  # Ensure we don't have weird spacing
+    favorite_ids = rest.split(" ")
+    multiple = len(favorite_ids) > 1
+    for favorite_id in favorite_ids:
+        favorite_global_id = IDS.to_global(favorite_id)
+        if favorite_global_id is None:
+            cprint(f"  Can't favorite id {favorite_id}: Not found", fg("red") + attr("bold"))
+            next
+        faved = mastodon.status_favourite(favorite_global_id)
+        msg = f"  Favorited ({favorite_id}):\n" + get_content(faved)
+        cprint(msg, attr("dim"))
+        if multiple:
+            print()
 
 
-@command("<id>", "Toots")
+@command("<id> [<id>]", "Toots")
 def unfav(mastodon, rest):
-    """Removes a favorite toot by ID."""
-    rest = IDS.to_global(rest)
-    if rest is None:
-        return
-    mastodon.status_unfavourite(rest)
-    unfaved = mastodon.status(rest)
-    msg = "  Removed favorite: " + get_content(unfaved)
-    cprint(msg, fg("yellow"))
+    """Removes a favorite toot by ID or IDs."""
+    for i in range(0, 2):
+        rest = rest.replace("  ", " ")  # Ensure we don't have weird spacing
+    favorite_ids = rest.split(" ")
+    multiple = len(favorite_ids) > 1
+    for favorite_id in favorite_ids:
+        favorite_global_id = IDS.to_global(favorite_id)
+        if favorite_global_id is None:
+            cprint(f"  Can't unfavorite id {favorite_id}: Not found", fg("red") + attr("bold"))
+            next
+        unfaved = mastodon.status_unfavourite(favorite_global_id)
+        msg = f"  Removed favorite ({favorite_id}):\n" + get_content(unfaved)
+        cprint(msg, fg("yellow"))
+        if multiple:
+            print()
 
 
 @command("<id>", "Toots")